+++ src/main/java/com/floney/floney/book/repository/category/CategoryRepositoryImpl.java
@@ -1,8 +1,8 @@
 package com.floney.floney.book.repository.category;
 
-import com.floney.floney.book.domain.category.Category;
 import com.floney.floney.book.domain.category.CategoryType;
-import com.floney.floney.book.domain.category.CustomSubCategory;
+import com.floney.floney.book.domain.category.entity.Category;
+import com.floney.floney.book.domain.category.entity.CustomSubCategory;
 import com.floney.floney.book.domain.entity.Book;
 import com.floney.floney.book.domain.entity.BookLine;
 import com.floney.floney.book.dto.process.CategoryInfo;
