+++ src/main/java/com/floney/floney/book/repository/category/CustomSubCategoryRepository.java
@@ -1,6 +1,6 @@
 package com.floney.floney.book.repository.category;
 
-import com.floney.floney.book.domain.category.CustomSubCategory;
+import com.floney.floney.book.domain.category.entity.CustomSubCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
